services/control-center/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2
websockets==12.0
psutil==5.9.6
pyyaml==6.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
